[cw-19]: add dict and metadata to contact us page
