+++ components/ContactUsPage/ContactUsHeader/ContactUsHeader.tsx
@@ -1,20 +1,23 @@
-import { Box, Typography } from "@mui/material"
+import { Box, Typography } from "@mui/material";
 
-import style from './ContactUsHeader.style'
+import { TFunction } from "i18next";
+import style from "./ContactUsHeader.style";
 
-const ContactUsHeader = () => {
+interface IContactUsHeader {
+  t: TFunction;
+}
+
+const ContactUsHeader = ({ t }: IContactUsHeader) => {
   return (
     <Box sx={style.container}>
       <Box sx={style.contentWrapper}>
-        <Typography sx={style.name}>
-          Contact Us
-        </Typography>
-        <Typography  variant="h5" sx={style.subtitle}>
-          Start the conversation to established good relationship and business.
+        <Typography sx={style.name}>{t("header.title")}</Typography>
+        <Typography variant="h5" sx={style.subtitle}>
+          {t("header.subtitle")}
         </Typography>
       </Box>
     </Box>
-  )
-}
+  );
+};
 
-export default ContactUsHeader
+export default ContactUsHeader;
